Refactored exisiting code for better coding purpose
