+++ src/components/PlatformListIcons.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Platform } from "../hooks/useGames";
+import { Platform } from "../hooks/usePlatforms";
 import {
   FaWindows,
   FaAndroid,
