+++ src/hooks/useData.ts
@@ -1,11 +1,7 @@
 import { useEffect, useState } from "react";
-import apiClient from "../services/api-client";
+import apiClient, { FetchResponse } from "../services/api-client";
 import { AxiosRequestConfig, CanceledError } from "axios";
 
-export interface FetchResponse<T> {
-  count: number;
-  results: T[];
-}
 const useData = <T>(
   urlPath: string,
   requestConfig?: AxiosRequestConfig,
