+++ src/hooks/useGenres.ts
@@ -1,8 +1,6 @@
 import genre from "../data/Genre";
 import { useQuery } from "@tanstack/react-query";
-import apiClient from "../services/api-client";
-import axios from "axios";
-import { FetchResponse } from "./useData";
+import apiClient, { FetchResponse } from "../services/api-client";
 
 export interface Genre {
   id: number;
