+++ src/hooks/usePlatforms.ts
@@ -1,18 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 import platforms from "../data/Platforms";
-import apiClient from "../services/api-client";
-import { FetchResponse } from "./useData";
+import apiClient, { FetchResponse } from "../services/api-client";
+
 export interface Platform {
   id: number;
   name: string;
   slug: string;
 }
-
-/*const usePlatforms = () => ({
-  data: platforms,
-  isLoading: false,
-  errors: null,
-});*/
 
 const usePlatforms = () =>
   useQuery({
@@ -21,6 +15,8 @@ const usePlatforms = () =>
       apiClient
         .get<FetchResponse<Platform>>("/platforms/lists/parents")
         .then((res) => res.data),
+    staleTime: 2 * 60 * 60 * 1000,
+    initialData: { count: platforms.length, results: platforms },
   });
 
 export default usePlatforms;
